+++ src/q_page.js
@@ -4,6 +4,7 @@
 var ensure = require("./util/ensure.js");
 var PageSize = require("./descriptors/page_size.js");
 var PageEdge = require("./descriptors/page_edge.js");
+var Center = require("./descriptors/center.js");
 
 var Me = module.exports = function QPage(frame) {
 	var QFrame = require("./q_frame.js");   // break circular dependency
@@ -17,4 +18,7 @@ var Me = module.exports = function QPage(frame) {
 	this.right = PageEdge.right(frame);
 	this.bottom = PageEdge.bottom(frame);
 	this.left = PageEdge.left(frame);
+
+	this.center = Center.x(this.left, this.right, "center of page");
+	this.middle = Center.y(this.top, this.bottom, "middle of page");
 };